refactor: usecase module 추출
