+++ usecase-module/src/main/java/org/c4marathon/assignment/settlement/SettlementUseCase.java
@@ -1,10 +1,10 @@
-package org.c4marathon.assignment.usecase;
+package org.c4marathon.assignment.settlement;
 
 import java.util.List;
 
-import org.c4marathon.assignment.domain.service.mainaccount.SettlementService;
-import org.c4marathon.assignment.dto.settlement.SettlementRequestDto;
-import org.c4marathon.assignment.dto.settlement.SettlementResponseDto;
+import org.c4marathon.assignment.domain.service.SettlementService;
+import org.c4marathon.assignment.settlement.dto.SettlementRequestDto;
+import org.c4marathon.assignment.settlement.dto.SettlementResponseDto;
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
